requirements.txt
aiohttp
pycrypto
requests
